Optimize register writes and color mode lookup
vbcc doesn't seem to optimize out unused variable assignments or attempt to make sense of bit shifts of 32-bit values. This optimization seems to result in a very large speedup, especially for 16 and 32 bit display modes.
The color mode lookup doesn't have as much of an impact, but it does save on quite a few function calls.